a little organization, check installs better

+++ mac/setup.sh
@@ -55,12 +55,15 @@ function masinstall {
 }
 
 function dmginstall {
-  appPath="$HOME/Applications/$1.app"
+  appPath="/Applications/$1.app"
+  appPathUser="$HOME/Applications/$1.app"
   downloadPath="$HOME/Downloads/$1.dmg"
-  if ! [ -d "$appPath" ] && askto "install $1"; then
+  # only offer to install if not installed in either the user or "all users" locations
+  if ! [ -d "$appPath" ] && ! [ -d "$appPathUser" ] && askto "install $1"; then
     curl -JL "$2" -o "$downloadPath"
     yes | hdiutil attach "$downloadPath" > /dev/null
-    ditto "/Volumes/$3/$1.app" "$appPath"
+    # install in the "all users" location
+    sudoit ditto "/Volumes/$3/$1.app" "$appPath"
     diskutil unmount "$3"
     rm "$downloadPath"
   fi
@@ -114,7 +117,7 @@ brews=(
   maven
   python
   python3
-  caskroom/cask/java
+  caskroom/cask/java # todo: current detection below doesn't see this, fix it
   shellcheck # shell script linting
   shpotify # Spotify shell CLI
   terraform
@@ -231,133 +234,159 @@ if ! [ -d "$HOME/Applications/iTerm.app" ]; then
   # todo: change plist directly for scroll back Root > New Bookmarks > Item 0 > Unlimited Scrollback > Boolean YES
 fi
 
-# Install OmniFocus 2
-# Beta builds may be available from - https://omnistaging.omnigroup.com/omnifocus-2/
-dmginstall "OmniFocus" https://www.omnigroup.com/download/latest/omnifocus/ "OmniFocus"
-# Install OmniGraffle
-dmginstall "OmniGraffle" https://downloads.omnigroup.com/software/MacOSX/10.11/OmniGraffle-7.3.1.dmg "OmniGraffle"
-# Install OmniOutliner 5
-# Beta builds may be available from - https://omnistaging.omnigroup.com/omnioutliner/
-dmginstall "OmniOutliner" https://www.omnigroup.com/download/latest/omnioutliner/ "OmniOutliner"
-# Install Google Chrome
-# disabled for now since Google Chrome does not seem to work correctly when in the user Applications folder.
-# dmginstall "Google Chrome" https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg "Google Chrome"
-# Install Spotify
-dmginstall "Spotify" https://download.spotify.com/Spotify.dmg "Spotify"
-# Install Docker
-dmginstall "Docker" https://download.docker.com/mac/stable/Docker.dmg "Docker"
-# Install Charles
-dmginstall "Charles" https://www.charlesproxy.com/assets/release/4.1.2/charles-proxy-4.1.2.dmg "Charles Proxy v4.1.2"
-# Install Skype
-dmginstall "Skype" https://get.skype.com/go/getskype-macosx "Skype"
-# Install Android Studio
-dmginstall "Android Studio" https://dl.google.com/dl/android/studio/install/2.3.1.0/android-studio-ide-162.3871768-mac.dmg "Android Studio 2.3.1"
-# Install XMind
-dmginstall "XMind" http://dl2.xmind.net/xmind-downloads/xmind-8-update2-macosx.dmg "XMind"
-# Install ControlPlane
-dmginstall "ControlPlane" https://www.dropbox.com/s/lhuyzp1csx3f9cc/ControlPlane-1.6.6.dmg?dl=1 "ControlPlane"
-# Install Blue Jeans Scheduler
-dmginstall "Blue Jeans Scheduler for Mac" https://swdl.bluejeans.com/bluejeansformac/Blue+Jeans+Scheduler+for+Mac-1.0.208.dmg "Blue Jeans Scheduler for Mac"
-# Install KeeWeb
-dmginstall "KeeWeb" https://github.com/keeweb/keeweb/releases/download/v1.5.4/KeeWeb-1.5.4.mac.dmg "KeeWeb"
-
 if ! type "aws" > /dev/null; then
+  echo "Installing AWS CLI..."
   # call sudoit to ensure password is set
-  sudoit echo "Installing AWS CLI..."
+  sudoit printf ""
   # passing -H instead of using sudoit
   # shellcheck disable=SC2002
   cat "$passfile" | sudo -H -S -p "" pip install awscli
 else
+  echo "Upgrading AWS CLI..."
   # call sudoit to ensure password is set
-  sudoit echo "Upgrading AWS CLI..."
+  sudoit printf ""
   # passing -H instead of using sudoit
   # shellcheck disable=SC2002
   cat "$passfile" | sudo -H -S -p "" pip install --upgrade awscli
 fi
 
-if askto "enable auto download & install of Mac App Store updates and macOS updates"; then
-  sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate AutomaticCheckEnabled -bool true
-  sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate AutomaticDownload -bool true
-  sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate CriticalUpdateInstall -bool true
-  sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate ConfigDataInstall -bool true
-  sudoit defaults write /Library/Preferences/com.apple.commerce AutoUpdate -bool true
-  sudoit defaults write /Library/Preferences/com.apple.commerce AutoUpdateRestartRequired -bool true
+if askto "review and install some recommended applications"; then
+  # Install OmniFocus 2
+  # Beta builds may be available from - https://omnistaging.omnigroup.com/omnifocus-2/
+  dmginstall "OmniFocus" https://www.omnigroup.com/download/latest/omnifocus/ "OmniFocus"
+  # Install OmniGraffle
+  dmginstall "OmniGraffle" https://www.omnigroup.com/download/latest/omnigraffle/ "OmniGraffle"
+  # Install OmniOutliner 5
+  # Beta builds may be available from - https://omnistaging.omnigroup.com/omnioutliner/
+  dmginstall "OmniOutliner" https://www.omnigroup.com/download/latest/omnioutliner/ "OmniOutliner"
+  # Install Google Chrome
+  dmginstall "Google Chrome" https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg "Google Chrome"
+  # Install Spotify
+  dmginstall "Spotify" https://download.spotify.com/Spotify.dmg "Spotify"
+  # Install Docker
+  dmginstall "Docker" https://download.docker.com/mac/stable/Docker.dmg "Docker"
+  # Install Charles
+  dmginstall "Charles" https://www.charlesproxy.com/assets/release/4.1.2/charles-proxy-4.1.2.dmg "Charles Proxy v4.1.2"
+  # Install Skype
+  dmginstall "Skype" https://get.skype.com/go/getskype-macosx "Skype"
+  # Install Android Studio
+  dmginstall "Android Studio" https://dl.google.com/dl/android/studio/install/2.3.1.0/android-studio-ide-162.3871768-mac.dmg "Android Studio 2.3.1"
+  # Install XMind
+  dmginstall "XMind" http://dl2.xmind.net/xmind-downloads/xmind-8-update2-macosx.dmg "XMind"
+  # Install ControlPlane
+  dmginstall "ControlPlane" https://www.dropbox.com/s/lhuyzp1csx3f9cc/ControlPlane-1.6.6.dmg?dl=1 "ControlPlane"
+  # Install Blue Jeans Scheduler
+  dmginstall "Blue Jeans Scheduler for Mac" https://swdl.bluejeans.com/bluejeansformac/Blue+Jeans+Scheduler+for+Mac-1.0.208.dmg "Blue Jeans Scheduler for Mac"
+  # Install KeeWeb
+  dmginstall "KeeWeb" https://github.com/keeweb/keeweb/releases/download/v1.5.4/KeeWeb-1.5.4.mac.dmg "KeeWeb"
+
+  # seeing if Excel is installed as crude check for Office
+  if ! [ -d "/Applications/Microsoft Excel.app" ] && askto "install Microsoft Office"; then
+    curl -JL https://go.microsoft.com/fwlink/?linkid=532572 -o "$HOME/Downloads/InstallOffice.pkg"
+    sudoit installer -pkg "$HOME/Downloads/InstallOffice.pkg" -target /
+    rm "$HOME/Downloads/InstallOffice.pkg"
+  fi
+
+  if ! [ -d "/Applications/Dropbox.app" ] && askto "install Dropbox"; then
+    dmg="$HOME/Downloads/Dropbox.dmg"
+    curl -JL https://www.dropbox.com/download?plat=mac -o "$dmg"
+    hdiutil attach "$HOME/Downloads/Dropbox.dmg"
+    open "/Volumes/Dropbox Installer/Dropbox.app"
+    read -r -p "Hit Enter when Dropbox has finished installing..."
+    diskutil unmount "Dropbox Installer"
+    rm "$dmg"
+  fi
+
+  if ! [ -d "$HOME/Applications/Blue Jeans.app" ] && askto "install Blue Jeans"; then
+    dmg="$HOME/Downloads/BlueJeans.dmg"
+    curl -JL https://swdl.bluejeans.com/desktop/mac/launchers/BlueJeansLauncher_live_168.dmg -o "$dmg"
+    hdiutil attach "$dmg"
+    open "/Volumes/Blue Jeans Launcher/Blue Jeans Launcher.app"
+    read -r -p "Hit Enter when Blue Jeans has finished installing..."
+    diskutil unmount "Blue Jeans Launcher"
+    rm "$dmg"
+  fi
+
+  if ! [ -d "/Applications/GPG Keychain.app" ] && ! askto "install GPG Suite"; then
+    dmg="$HOME/Downloads/GPGSuite.dmg"
+    curl -JL https://releases.gpgtools.org/GPG_Suite-2017.1b3-v2.dmg -o "$dmg"
+    hdiutil attach "$dmg"
+    sudoit installer -pkg "/Volumes/GPG Suite/Install.pkg" -target /
+    diskutil unmount "GPG Suite"
+    rm "$dmg"
+  fi
+
+  # Install YubiKey PIV Manager to enable unlock with a YubiKey
+  if ! [ -d "/Applications/YubiKey PIV Manager.app" ] && askto "install YubiKey PIV Manager"; then
+    pkg="$HOME/Downloads/YubiKeyPIVManager.pkg"
+    curl -JL https://developers.yubico.com/yubikey-piv-manager/Releases/yubikey-piv-manager-1.4.1-mac.pkg -o "$pkg"
+    sudoit installer -pkg "$pkg" -target /
+    rm "$pkg"
+  fi
+
+  if ! [ -d "/Applications/Keyboard Maestro.app" ] && askto "install Keyboard Maestro"; then
+    curl -JL https://files.stairways.com/keyboardmaestro-731.zip -o "$HOME/Downloads/KeyboardMaestro.zip"
+    unzip -q "$HOME/Downloads/KeyboardMaestro.zip" -d "/Applications"
+    rm "$HOME/Downloads/KeyboardMaestro.zip"
+  fi
+
+  if ! [ -d "$HOME/Applications/Atom.app" ] && askto "install Atom"; then
+    curl -JL https://atom.io/download/mac -o "$HOME/Downloads/Atom.zip"
+    unzip -q "$HOME/Downloads/Atom.zip" -d "$HOME/Applications"
+    rm "$HOME/Downloads/Atom.zip"
+  fi
+
+  # Ensure Atom Shell Commands are installed
+  if [ -d "$HOME/Applications/Atom.app" ] && ! type "apm" > /dev/null; then
+    echo "Please install the Atom shell commands from inside Atom."
+    echo "From the Atom menu bar, select Atom > Install Shell Commands."
+    read -r -p "Hit Enter to open Atom..."
+    open "$HOME/Applications/Atom.app"
+    read -r -p "Hit Enter when finished installing the shell commands..."
+    if ! apm list | grep "── vim-mode@" > /dev/null && askto "install Atom vim-mode"; then
+      apm install vim-mode ex-mode
+    fi
+  fi
+
+  if ! [ -d "$HOME/Applications/Atom Beta.app" ] && askto "install Atom Beta"; then
+    curl -JL https://atom.io/download/mac?channel=beta -o "$HOME/Downloads/AtomBeta.zip"
+    unzip -q "$HOME/Downloads/AtomBeta.zip" -d "$HOME/Applications"
+    rm "$HOME/Downloads/AtomBeta.zip"
+  fi
+
+  if ! [ -d "$HOME/Applications/Visual Studio Code.app" ] && askto "install Visual Studio Code"; then
+    curl -JL https://go.microsoft.com/fwlink/?LinkID=620882 -o "$HOME/Downloads/VSCode.zip"
+    unzip -q "$HOME/Downloads/VSCode.zip" -d "$HOME/Applications/"
+    rm "$HOME/Downloads/VSCode.zip"
+  fi
+
+  if ! [ -d "$HOME/Applications/Visual Studio Code - Insiders.app" ] && askto "install Visual Studio Code Insiders"; then
+    curl -JL https://go.microsoft.com/fwlink/?LinkId=723966 -o "$HOME/Downloads/VSCodeInsiders.zip"
+    unzip -q "$HOME/Downloads/VSCodeInsiders.zip" -d "$HOME/Applications/"
+    rm "$HOME/Downloads/VSCodeInsiders.zip"
+  fi
+
+  if ! [ -d "$HOME/Applications/Beyond Compare.app" ] && askto "install Beyond Compare"; then
+    curl -JL http://www.scootersoftware.com/BCompareOSX-4.2.2.22384.zip -o "$HOME/Downloads/BeyondCompare.zip"
+    unzip -q "$HOME/Downloads/BeyondCompare.zip" -d "$HOME/Applications/"
+    rm "$HOME/Downloads/BeyondCompare.zip"
+  fi
+
 fi
 
-# Install Office 365
-if ! [ -d "/Applications/Microsoft Excel.app" ] && askto "install Microsoft Office"; then
-  curl -JL https://go.microsoft.com/fwlink/?linkid=532572 -o "$HOME/Downloads/InstallOffice.pkg"
-  sudoit installer -pkg "$HOME/Downloads/InstallOffice.pkg" -target /
-  rm "$HOME/Downloads/InstallOffice.pkg"
-fi
-
-# Install Dropbox https://www.dropbox.com/download?plat=mac
-if ! [ -d "/Applications/Dropbox.app" ] && askto "install Dropbox"; then
-  dmg="$HOME/Downloads/Dropbox.dmg"
-  curl -JL https://www.dropbox.com/download?plat=mac -o "$dmg"
-  hdiutil attach "$HOME/Downloads/Dropbox.dmg"
-  open "/Volumes/Dropbox Installer/Dropbox.app"
-  read -r -p "Hit Enter when Dropbox has finished installing..."
-  diskutil unmount "Dropbox Installer"
-  rm "$dmg"
-fi
-
-# Install Blue Jeans Launcher
-if ! [ -d "$HOME/Applications/Blue Jeans.app" ] && askto "install Blue Jeans"; then
-  dmg="$HOME/Downloads/BlueJeans.dmg"
-  curl -JL https://swdl.bluejeans.com/desktop/mac/launchers/BlueJeansLauncher_live_168.dmg -o "$dmg"
-  hdiutil attach "$dmg"
-  open "/Volumes/Blue Jeans Launcher/Blue Jeans Launcher.app"
-  read -r -p "Hit Enter when Blue Jeans has finished installing..."
-  diskutil unmount "Blue Jeans Launcher"
-  rm "$dmg"
-fi
-
-# Install GPG Suite
-if ! [ -d "/Applications/GPG Keychain.app" ] && ! askto "install GPG Suite"; then
-  dmg="$HOME/Downloads/GPGSuite.dmg"
-  curl -JL https://releases.gpgtools.org/GPG_Suite-2017.1b3-v2.dmg -o "$dmg"
-  hdiutil attach "$dmg"
-  sudoit installer -pkg "/Volumes/GPG Suite/Install.pkg" -target /
-  diskutil unmount "GPG Suite"
-  rm "$dmg"
-fi
-
-# Install YubiKey PIV Manager to enable unlock with a YubiKey
-if ! [ -d "/Applications/YubiKey PIV Manager.app" ] && askto "install YubiKey PIV Manager"; then
-  pkg="$HOME/Downloads/YubiKeyPIVManager.pkg"
-  curl -JL https://developers.yubico.com/yubikey-piv-manager/Releases/yubikey-piv-manager-1.4.1-mac.pkg -o "$pkg"
-  sudoit installer -pkg "$pkg" -target /
-  rm "$pkg"
-fi
-
-# Keyboard Maestro
-if ! [ -d "/Applications/Keyboard Maestro.app" ] && askto "install Keyboard Maestro"; then
-  curl -JL https://files.stairways.com/keyboardmaestro-731.zip -o "$HOME/Downloads/KeyboardMaestro.zip"
-  unzip -q "$HOME/Downloads/KeyboardMaestro.zip" -d "/Applications"
-  rm "$HOME/Downloads/KeyboardMaestro.zip"
-fi
-
-# Install Atom
-if ! [ -d "$HOME/Applications/Atom.app" ] && askto "install Atom"; then
-  curl -JL https://atom.io/download/mac -o "$HOME/Downloads/Atom.zip"
-  unzip -q "$HOME/Downloads/Atom.zip" -d "$HOME/Applications"
-  rm "$HOME/Downloads/Atom.zip"
-fi
-
-if [ -d "$HOME/Applications/Atom.app" ] && ! type "apm" > /dev/null; then
-  echo "Please install the Atom shell commands from inside Atom."
-  echo "From the Atom menu bar, select Atom > Install Shell Commands."
-  read -r -p "Hit Enter to open Atom..."
-  open "$HOME/Applications/Atom.app"
-  read -r -p "Hit Enter when finished installing the shell commands..."
-fi
-
-# Install Atom Beta
-if ! [ -d "$HOME/Applications/Atom Beta.app" ] && askto "install Atom Beta"; then
-  curl -JL https://atom.io/download/mac?channel=beta -o "$HOME/Downloads/AtomBeta.zip"
-  unzip -q "$HOME/Downloads/AtomBeta.zip" -d "$HOME/Applications"
-  rm "$HOME/Downloads/AtomBeta.zip"
+if ! (defaults read /Library/Preferences/com.apple.SoftwareUpdate AutomaticCheckEnabled && \
+  defaults read /Library/Preferences/com.apple.SoftwareUpdate AutomaticDownload && \
+  defaults read /Library/Preferences/com.apple.SoftwareUpdate CriticalUpdateInstall && \
+  defaults read /Library/Preferences/com.apple.SoftwareUpdate ConfigDataInstall && \
+  defaults read /Library/Preferences/com.apple.commerce AutoUpdate && \
+  defaults read /Library/Preferences/com.apple.commerce AutoUpdateRestartRequired) > /dev/null && \
+  askto "enable auto download & install of Mac App Store updates and macOS updates"; then
+    sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate AutomaticCheckEnabled -bool true
+    sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate AutomaticDownload -bool true
+    sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate CriticalUpdateInstall -bool true
+    sudoit defaults write /Library/Preferences/com.apple.SoftwareUpdate ConfigDataInstall -bool true
+    sudoit defaults write /Library/Preferences/com.apple.commerce AutoUpdate -bool true
+    sudoit defaults write /Library/Preferences/com.apple.commerce AutoUpdateRestartRequired -bool true
 fi
 
 # Install atom packages
@@ -402,28 +431,6 @@ if type "apm" > /dev/null; then
     fi
   done
   rm "$apmtempfile"
-  if ! apm list | grep "── vim-mode@" > /dev/null && askto "install Atom vim-mode"; then
-    apm install vim-mode ex-mode
-  fi
-fi
-
-
-if ! [ -d "$HOME/Applications/Visual Studio Code.app" ] && askto "install Visual Studio Code"; then
-  curl -JL https://go.microsoft.com/fwlink/?LinkID=620882 -o "$HOME/Downloads/VSCode.zip"
-  unzip -q "$HOME/Downloads/VSCode.zip" -d "$HOME/Applications/"
-  rm "$HOME/Downloads/VSCode.zip"
-fi
-
-if ! [ -d "$HOME/Applications/Visual Studio Code - Insiders.app" ] && askto "install Visual Studio Code Insiders"; then
-  curl -JL https://go.microsoft.com/fwlink/?LinkId=723966 -o "$HOME/Downloads/VSCodeInsiders.zip"
-  unzip -q "$HOME/Downloads/VSCodeInsiders.zip" -d "$HOME/Applications/"
-  rm "$HOME/Downloads/VSCodeInsiders.zip"
-fi
-
-if ! [ -d "$HOME/Applications/Beyond Compare.app" ] && askto "install Beyond Compare"; then
-  curl -JL http://www.scootersoftware.com/BCompareOSX-4.2.2.22384.zip -o "$HOME/Downloads/BeyondCompare.zip"
-  unzip -q "$HOME/Downloads/BeyondCompare.zip" -d "$HOME/Applications/"
-  rm "$HOME/Downloads/BeyondCompare.zip"
 fi
 
 # Quick Look Generators
@@ -432,9 +439,11 @@ if ! [ -d /Library/QuickLook/Provisioning.qlgenerator ]; then
   curl -JL "https://github.com/chockenberry/Provisioning/releases/download/1.0.4/Provisioning-1.0.4.zip" -o "$HOME/Downloads/qlprovisioning.zip"
   unzip -q "$HOME/Downloads/qlprovisioning.zip"
   sudoit mv "$HOME/Downloads/Provisioning-1.0.4/Provisioning.qlgenerator" /Library/QuickLook
+  rm "$HOME/Downloads/qlprovisioning.zip"
+  rm -rf "$HOME/Downloads/Provisioning-1.0.4/"
 fi
 
-if askto "set some starter system settings"; then
+if askto "set some opinionated starter system settings"; then
   echo "Modifying System Settings"
   echo "Only show icons of running apps in app bar, using Spotlight to launch"
   defaults write com.apple.dock static-only -bool true
@@ -691,14 +700,14 @@ if askto "set some starter system settings"; then
 fi
 
 GITHUB_KEYS_URL="https://github.com/settings/keys"
-if askto "configure git settings"; then
+if askto "configure git, create an SSH key for GitHub or a GPG key to sign commits"; then
   read -r -p "Enter your GitHub email address: " GITHUB_EMAIL
   read -r -p "Enter your full name used on GitHub: " FULL_NAME
   # configure git
   git config --global user.name "$FULL_NAME"
   git config --global user.email "$GITHUB_EMAIL"
   # Generate a new SSH key for GitHub https://help.github.com/articles/generating-a-new-ssh-key-and-adding-it-to-the-ssh-agent/
-  if askto "create a GitHub ssh key for $GITHUB_EMAIL"; then
+  if askto "create a GitHub SSH key for $GITHUB_EMAIL"; then
     ssh-keygen -t rsa -b 4096 -C "$GITHUB_EMAIL"
     # start ssh-agent
     eval "$(ssh-agent -s)"
@@ -755,10 +764,11 @@ EOF
 fi
 
 if false; then
-# todo: Install JDK 8
-open http://www.oracle.com/technetwork/java/javase/downloads/jdk8-downloads-2133151.html
 # todo: setup ControlPlane
 open "$HOME/Applications/ControlPlane.app"
+# todo: install Adobe Connect
+https://www.adobe.com/go/adobeconnect_9_addin_mac
+# adobeconnectaddin-installer.pkg
 # todo: Install CoRD
 # todo: Install Sketch
 # todo: Install SourceTree
